requirements.txt
# requirements.txt
fastapi
uvicorn
redis
pydantic
passlib[bcrypt]
python-multipart
python-jose[cryptography]
Jinja2
pydantic[email]